+++ ethereum-analyzer/src/Ethereum/Analyzer/Solidity/Hoople.hs
@@ -9,8 +9,10 @@ module Ethereum.Analyzer.Solidity.Hoople
 import Protolude hiding ((<*>), show)
 
 import Compiler.Hoopl
-import GHC.Show (Show(..))
 import Ethereum.Analyzer.Solidity.Simple
+import GHC.Show (Show(..))
+import Text.PrettyPrint.Leijen.Text hiding ((<$>))
+import qualified Text.PrettyPrint.Leijen.Text as PP
 
 data HContract = HContract
   { hcName :: Text
@@ -29,19 +31,27 @@ data HFunDefinition = HFunDefinition
   , hfCFG :: CFG
   }
 
+instance Pretty Label where
+  pretty = textStrict . toS . show
+
 data HStatement e x where
-        CoSt :: Label -> HStatement C O
-        OoSt :: Statement -> HStatement O O
-        OcSt :: Statement -> [Label] -> HStatement O C
-        OcJump :: Label -> HStatement O C
+  CoSt :: Label -> HStatement C O
+  OoSt :: Statement -> HStatement O O
+  OcSt :: Statement -> [Label] -> HStatement O C
+  OcJump :: Label -> HStatement O C
+
+instance Pretty (HStatement e x) where
+  pretty (CoSt l) = pretty l
+  pretty (OoSt st) = pretty st
+  pretty (OcSt (StIf lval _ _) ll) =
+    textStrict "OcIf" <+> pretty lval <+> textStrict "->" <+> prettyList ll
+  pretty (OcSt (StLoop _) ll) =
+    textStrict "OcLoop" <+> textStrict "->" <+> prettyList ll
+  pretty (OcSt st ll) = pretty st <+> textStrict "->" <+> prettyList ll
+  pretty (OcJump l) = textStrict "jump" <+> textStrict "->" <+> pretty l
 
 instance Show (HStatement e x) where
-  show (CoSt l) = show l
-  show (OoSt st) = show st
-  show (OcSt (StIf lval _ _) ll) = "if " <> show lval <> " -> " <> show ll
-  show (OcSt (StLoop _) ll) = "loop -> " <> show ll
-  show (OcSt st ll) = show st <> " -> " <> show ll
-  show (OcJump l) = "jump -> " <> show l
+  show = show . pretty
 
 instance NonLocal HStatement where
   entryLabel (CoSt lbl) = lbl
@@ -52,16 +62,12 @@ instance HooplNode HStatement where
   mkBranchNode = OcJump
   mkLabelNode = CoSt
 
-hoopleOf
-  :: UniqueMonad m
-  => Contract -> m HContract
+hoopleOf :: UniqueMonad m => Contract -> m HContract
 hoopleOf Contract {cName = name, cStateVars = vars, cFunctions = funs} = do
   hFuns <- mapM hfunOf funs
   return $ HContract name vars hFuns
 
-hfunOf
-  :: UniqueMonad m
-  => FunDefinition -> m HFunDefinition
+hfunOf :: UniqueMonad m => FunDefinition -> m HFunDefinition
 hfunOf FunDefinition { fName = name
                      , fParams = params
                      , fReturns = returns
@@ -74,9 +80,7 @@ hfunOf FunDefinition { fName = name
   cfg <- graphOfAGraph acfg
   return $ HFunDefinition name params returns cfg
 
-graphOf
-  :: UniqueMonad m
-  => [Statement] -> Label -> Label -> m ACFG
+graphOf :: UniqueMonad m => [Statement] -> Label -> Label -> m ACFG
 graphOf [] exitL _ = return $ mkBranch exitL
 graphOf (h:t) exitL loopL = do
   postDom <- graphOf t exitL loopL